feat: rework module system
Now the module system will return a module key instead of a bool, this
allows for a module to pick any other module to move onto.

there is a pseudo module quit which will push a quit event. then each
module will get a key in the module enum in types.lua
the next module is set to the current on the finish of update and we
keep turning