feat(lib.button): finish basic button class
Button:new() takes in a table of arguments, this allows us to have the
illusion of named arguments which makes the construction easier to read.

There is now a Button:setPosition() method.

A button will only call its callback when Button:inBound() is true

Button:draw() will now draw the buttons text in about the center